Refactor stats calculation
