+++ src/app/pages/Explore/BakingHistoryItem.tsx
@@ -20,7 +20,7 @@ import HashChip from "app/templates/HashChip";
 import { getPluralKey, toLocalFormat } from "lib/i18n/numbers";
 import { T } from "lib/i18n/react";
 import {
-  defaultRewardConfigHistory,
+  getRewardsStats,
   mutezToTz,
   useExplorerBaseUrls,
   useKnownBaker,
@@ -51,27 +51,12 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
   const {
     cycle,
     baker,
-    balance,
     ownBlockRewards,
-    extraBlockRewards,
-    futureBlockRewards,
     endorsementRewards,
-    futureEndorsementRewards,
-    stakingBalance,
-    expectedBlocks,
-    expectedEndorsements,
     ownBlocks,
-    futureBlocks,
-    futureEndorsements,
     endorsements,
     ownBlockFees,
-    extraBlockFees,
-    revelationRewards,
-    doubleBakingRewards,
-    doubleEndorsingRewards,
     missedEndorsementRewards,
-    missedExtraBlockRewards,
-    missedExtraBlockFees,
     missedOwnBlockFees,
     missedOwnBlockRewards,
     missedOwnBlocks,
@@ -88,45 +73,39 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
   );
 
   const { StatusIcon, iconColor, title, statsEntriesProps } = useMemo(() => {
-    let rewardConfig = defaultRewardConfigHistory[0].value;
-    if (bakerDetails?.rewardConfigHistory) {
-      const { rewardConfigHistory } = bakerDetails;
-      for (let i = 0; i < rewardConfigHistory.length; i++) {
-        const historyEntry = rewardConfigHistory[i];
-        if (cycle >= historyEntry.cycle) {
-          rewardConfig = historyEntry.value;
-          break;
-        }
-      }
-    }
+    const {
+      balance,
+      rewards,
+      luck,
+      bakerFeePart,
+      bakerFee,
+      cycleStatus,
+      efficiency,
+    } = getRewardsStats({
+      rewardsEntry: content,
+      bakerDetails,
+      currentCycle,
+      fallbackRewardPerEndorsement,
+      fallbackRewardPerFutureBlock,
+      fallbackRewardPerFutureEndorsement,
+      fallbackRewardPerOwnBlock,
+    });
 
-    const totalFutureRewards = new BigNumber(futureEndorsementRewards).plus(
-      futureBlockRewards
-    );
-    const totalCurrentRewards = new BigNumber(extraBlockRewards)
-      .plus(endorsementRewards)
-      .plus(ownBlockRewards)
-      .plus(ownBlockFees)
-      .plus(extraBlockFees)
-      .plus(revelationRewards)
-      .plus(doubleBakingRewards)
-      .plus(doubleEndorsingRewards);
     const { StatusIcon, iconColor, title } = (() => {
-      switch (true) {
-        case totalFutureRewards.eq(0) &&
-          (currentCycle === undefined || cycle <= currentCycle - 6):
+      switch (cycleStatus) {
+        case "unlocked":
           return {
             StatusIcon: OkIcon,
             iconColor: "green-500",
             title: "Rewards unlocked",
           };
-        case totalFutureRewards.eq(0):
+        case "locked":
           return {
             StatusIcon: TimeIcon,
             iconColor: "orange-500",
             title: "Rewards still locked",
           };
-        case totalCurrentRewards.eq(0):
+        case "future":
           return {
             StatusIcon: HourglassIcon,
             iconColor: "gray-500",
@@ -140,67 +119,7 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
           };
       }
     })();
-    const totalRewards = totalFutureRewards.plus(totalCurrentRewards);
-    const rewards = totalRewards.multipliedBy(balance).div(stakingBalance);
-    let luckPercentage =
-      expectedBlocks + expectedEndorsements > 0
-        ? new BigNumber(-100)
-        : new BigNumber(0);
-    if (totalFutureRewards.plus(totalCurrentRewards).gt(0)) {
-      const rewardPerOwnBlock =
-        ownBlocks === 0
-          ? fallbackRewardPerOwnBlock
-          : new BigNumber(ownBlockRewards).div(ownBlocks);
-      const rewardPerEndorsement =
-        endorsements === 0
-          ? fallbackRewardPerEndorsement
-          : new BigNumber(endorsementRewards).div(endorsements);
-      const asIfNoFutureExpectedBlockRewards = new BigNumber(
-        expectedBlocks
-      ).multipliedBy(rewardPerOwnBlock);
-      const asIfNoFutureExpectedEndorsementRewards = new BigNumber(
-        expectedEndorsements
-      ).multipliedBy(rewardPerEndorsement);
-      const asIfNoFutureExpectedRewards = asIfNoFutureExpectedBlockRewards.plus(
-        asIfNoFutureExpectedEndorsementRewards
-      );
-
-      const rewardPerFutureBlock =
-        futureBlocks === 0
-          ? fallbackRewardPerFutureBlock
-          : new BigNumber(futureBlockRewards).div(futureBlocks);
-      const rewardPerFutureEndorsement =
-        futureEndorsements === 0
-          ? fallbackRewardPerFutureEndorsement
-          : new BigNumber(futureEndorsementRewards).div(futureEndorsements);
-      const asIfNoCurrentExpectedBlockRewards = new BigNumber(
-        expectedBlocks
-      ).multipliedBy(rewardPerFutureBlock);
-      const asIfNoCurrentExpectedEndorsementRewards = new BigNumber(
-        expectedEndorsements
-      ).multipliedBy(rewardPerFutureEndorsement);
-      const asIfNoCurrentExpectedRewards = asIfNoCurrentExpectedBlockRewards.plus(
-        asIfNoCurrentExpectedEndorsementRewards
-      );
-
-      const weights =
-        endorsements + futureEndorsements === 0
-          ? { current: ownBlocks, future: futureBlocks }
-          : { current: endorsements, future: futureEndorsements };
-      const totalExpectedRewards =
-        weights.current + weights.future === 0
-          ? new BigNumber(0)
-          : asIfNoFutureExpectedRewards
-              .multipliedBy(weights.current)
-              .plus(asIfNoCurrentExpectedRewards.multipliedBy(weights.future))
-              .div(new BigNumber(weights.current).plus(weights.future));
-
-      luckPercentage = totalRewards
-        .minus(totalExpectedRewards)
-        .div(totalExpectedRewards)
-        .multipliedBy(100)
-        .decimalPlaces(0);
-    }
+    const luckPercentage = luck.times(100);
     const luckClassName = (() => {
       switch (true) {
         case luckPercentage.lt(-5):
@@ -213,63 +132,11 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
     })();
     const normalizedBalance = mutezToTz(balance);
     const normalizedRewards = mutezToTz(rewards);
-    let bakerFeePart = bakerDetails?.fee ?? 0;
-    if (bakerDetails?.feeHistory) {
-      const { feeHistory } = bakerDetails;
-      for (let i = 0; i < feeHistory.length; i++) {
-        const historyEntry = feeHistory[i];
-        if (cycle >= historyEntry.cycle) {
-          bakerFeePart = historyEntry.value;
-          break;
-        }
-      }
-    }
-    const bakerFee = normalizedRewards
-      .multipliedBy(bakerFeePart)
-      .decimalPlaces(6);
-
-    const totalFutureRewards2 = new BigNumber(
-      rewardConfig.endorses ? futureEndorsementRewards : 0
-    ).plus(rewardConfig.blocks ? futureBlockRewards : 0);
-    const totalCurrentRewards2 = new BigNumber(
-      rewardConfig.blocks
-        ? new BigNumber(extraBlockRewards).plus(ownBlockRewards)
-        : 0
-    )
-      .plus(
-        rewardConfig.endorses
-          ? new BigNumber(endorsementRewards).plus(doubleEndorsingRewards)
-          : 0
-      )
-      .plus(
-        rewardConfig.fees ? new BigNumber(ownBlockFees).plus(extraBlockFees) : 0
-      )
-      .plus(rewardConfig.revelationRewards ? revelationRewards : 0)
-      .plus(doubleBakingRewards);
-    const totalRewards2 = totalFutureRewards2.plus(totalCurrentRewards2);
-
-    const fullEfficiencyIncome = new BigNumber(4e7)
-      .multipliedBy(new BigNumber(ownBlocks).plus(futureBlocks))
-      .plus(
-        new BigNumber(1.25e6).multipliedBy(
-          new BigNumber(endorsements).plus(futureEndorsements)
-        )
-      );
-    const totalLost = new BigNumber(missedEndorsementRewards)
-      .plus(missedExtraBlockFees)
-      .plus(missedExtraBlockRewards)
-      .plus(missedOwnBlockFees)
-      .plus(missedOwnBlockRewards);
-    const totalGain = totalRewards2
-      .minus(totalLost)
-      .minus(fullEfficiencyIncome);
-    const efficiencyPercentage = new BigNumber(1)
-      .plus(totalGain.div(fullEfficiencyIncome))
-      .multipliedBy(100)
-      .decimalPlaces(2);
+    const normalizedBakerFee = mutezToTz(bakerFee);
+    const efficiencyPercentage = efficiency.multipliedBy(100);
     const efficiencyClassName = (() => {
       switch (true) {
-        case totalFutureRewards.gt(0) && totalCurrentRewards.gt(0):
+        case cycleStatus === "inProgress":
           return "text-blue-600";
         case efficiencyPercentage.gte(100):
           return "text-green-500";
@@ -293,7 +160,7 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
                 "<1"
               ) : (
                 <Money smallFractionFont={false}>
-                  {mutezToTz(balance).decimalPlaces(0)}
+                  {mutezToTz(balance).decimalPlaces(0, BigNumber.ROUND_FLOOR)}
                 </Money>
               )}{" "}
               ꜩ
@@ -319,67 +186,46 @@ const BakingHistoryItem: FC<BakingHistoryItemProps> = ({
           value: `${bakerFeePart * 100}%`,
           valueComment: (
             <span className="text-gray-500">
-              (<Money smallFractionFont={false}>{bakerFee}</Money> ꜩ)
+              (<Money smallFractionFont={false}>{normalizedBakerFee}</Money> ꜩ)
             </span>
           ),
         },
         {
           name: "Expected payout",
-          value: totalCurrentRewards.eq(0) ? (
-            "‒"
-          ) : (
-            <>
-              <Money smallFractionFont={false}>
-                {normalizedRewards.minus(bakerFee)}
-              </Money>{" "}
-              ꜩ
-            </>
-          ),
+          value:
+            cycleStatus === "future" ? (
+              "‒"
+            ) : (
+              <>
+                <Money smallFractionFont={false}>
+                  {normalizedRewards.minus(normalizedBakerFee)}
+                </Money>{" "}
+                ꜩ
+              </>
+            ),
         },
         {
           className: "pb-0",
           name: "Efficiency",
-          value: totalCurrentRewards.eq(0) ? (
-            "‒"
-          ) : (
-            <span className={efficiencyClassName}>
-              {toLocalFormat(efficiencyPercentage, {})}%
-            </span>
-          ),
+          value:
+            cycleStatus === "future" ? (
+              "‒"
+            ) : (
+              <span className={efficiencyClassName}>
+                {toLocalFormat(efficiencyPercentage, { decimalPlaces: 2 })}%
+              </span>
+            ),
         },
       ],
     };
   }, [
-    balance,
-    currentCycle,
-    cycle,
-    ownBlockRewards,
-    extraBlockRewards,
-    futureBlockRewards,
-    endorsementRewards,
-    futureEndorsementRewards,
-    stakingBalance,
-    expectedBlocks,
-    expectedEndorsements,
-    ownBlocks,
-    futureBlocks,
-    futureEndorsements,
-    endorsements,
-    ownBlockFees,
-    extraBlockFees,
     bakerDetails,
+    content,
+    currentCycle,
     fallbackRewardPerEndorsement,
     fallbackRewardPerFutureBlock,
     fallbackRewardPerFutureEndorsement,
     fallbackRewardPerOwnBlock,
-    revelationRewards,
-    doubleBakingRewards,
-    doubleEndorsingRewards,
-    missedEndorsementRewards,
-    missedExtraBlockRewards,
-    missedExtraBlockFees,
-    missedOwnBlockFees,
-    missedOwnBlockRewards,
   ]);
 
   const accordionItemsProps = useMemo<AccordionItemProps[]>(
